move site url to compile time api
